Added uploading cancel to the MainModel.java

+++ src/main/java/com/example/twinfileshare/fx/model/MainModel.java
@@ -4,6 +4,7 @@ import com.example.twinfileshare.repository.GoogleUserCREDRepository;
 import com.example.twinfileshare.service.GoogleAuthorizationService;
 import com.example.twinfileshare.service.GoogleDriveService;
 import com.example.twinfileshare.service.utility.Zipper;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.scheduling.annotation.Async;
@@ -15,6 +16,7 @@ import java.security.GeneralSecurityException;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
+@Log4j2
 @Service
 public class MainModel {
 
@@ -36,7 +38,7 @@ public class MainModel {
         googleUserCREDRepository.deleteByEmail(email);
     }
 
-    private boolean isUploadCancelled;
+    private boolean isUploadingActive;
 
     @Autowired
     private GoogleDriveService driveService;
@@ -49,13 +51,20 @@ public class MainModel {
 
     @Async
     public CompletableFuture<Boolean> uploadFilesToGoogleDrive(String email, List<File> requiredFiles,
-                                                               String zipName) throws IOException, InterruptedException, GeneralSecurityException {
+                                                               String zipName) throws IOException {
+        if (isUploadingActive)
+            throw new IllegalStateException("An upload is active. Cancel the previous upload to start new.");
+
         System.out.println("Uploading to google drive account: " + email);
 
         zipper.zipFiles(requiredFiles, zipName);
 
+        isUploadingActive = true;
+
         var file = new File(zipName + ".zip");
         driveService.uploadFile(email, file);
+
+        isUploadingActive = false;
 
         var isUploadedLocalFileDeleted = file.delete();
         if (isUploadedLocalFileDeleted)
@@ -66,14 +75,26 @@ public class MainModel {
 
     @Async
     public CompletableFuture<Boolean> uploadFileToGoogleDrive(String email, File file) throws IOException, InterruptedException, GeneralSecurityException {
+        if (isUploadingActive)
+            throw new IllegalStateException("An upload is active. Cancel the previous upload to start new.");
+
         System.out.println("Uploading to google drive account: " + email);
 
+        isUploadingActive = true;
+
         driveService.uploadFile(email, file);
+
+        isUploadingActive = false;
 
         return CompletableFuture.completedFuture(true);
     }
 
     public void cancelUploadFiles() {
-        isUploadCancelled = true;
+        if (!isUploadingActive) {
+            driveService.cancelUpload();
+            return;
+        }
+
+       log.error("No upload to cancel.");
     }
 }
